added card component
