+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import AddUser from "./components/UI/Users/AddUser";
+import AddUser from "./components/Users/AddUser";
 
 function App() {
   return (
