+++ src/components/Users/AddUser.js
@@ -1,0 +1,23 @@
+import Card from "../UI/Card";
+import classes from "./AddUser.module.css";
+
+const AddUser = () => {
+  const addUserHandler = (e) => {
+    e.preventDefault();
+    console.log(e);
+  };
+
+  return (
+    <Card className={classes.input}>
+      <form onSubmit={addUserHandler}>
+        <label htmlFor="username">UserName</label>
+        <input id="username" type="text"></input>
+        <label htmlFor="age">Age (Years)</label>
+        <input id="age" type="number"></input>
+        <button type="submit">Add User</button>
+      </form>
+    </Card>
+  );
+};
+
+export default AddUser;
