+++ src/index.css
@@ -4,6 +4,7 @@
 
 html {
   font-family: sans-serif;
+  background-color: #1f1f1f;
 }
 
 body {
